makefile: make cpp examples compilable
Note that g++ can be able to compile and link the cpp files on darwin
but on fedora it fails to link because of compiling together c and c++
files. It's just a small fix. The main problem (to be able to compile by
both: g++ and clang++) will be fixed asap.